pyproject.toml
[tool.poetry]
name = "instagram-django-server"
version = "0.1.0"
description = ""
authors = ["ramram98 <tnfkachzh@naver.com>"]
readme = "README.md"
packages = [{include = "instagram_django_server"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.7"
selenium = "^4.8.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
